Ouali
reglage de bug avec morgan le fait d'attaquer trop loin dans sa zone
et dynamisation des animations

+++ Assets/Script/StaminaBar.cs
@@ -11,7 +11,7 @@ public class StaminaBar : MonoBehaviour
         slider.value = Stamina;
     }
 
-    public void SetStamina(int stamina)
+    public void SetStamina(float stamina)
 {
     slider.value = stamina; // Assure que le slider est bien mis à jour
     // Debug.Log($"Endurance mise à jour sur la barre : {stamina}");
